Avoid stopping propogation of product clicks
Stopping the propogation results in customers getting a surprise
when using cmd+click or equivalent to open the link in new window.

The beacon transport is supported only in some browsers, but it
is not worth it to let customers suffer for the sake of analytics.

About beacon: https://developers.google.com/analytics/devguides/collection/analyticsjs/field-reference#transport